+++ modules/home/default.nix
@@ -1,0 +1,8 @@
+{
+  home-manager = {
+    useGlobalPkgs = true;
+    useUserPackages = true;
+    backupFileExtension = "hm-bak";
+    users.avivaldelli = import ../../home/avivaldelli;
+  };
+}